ENH: Add DailyReturns builtin.
- Moved many non-technical analysis factors from
  `zipline.pipeline.factors.technical` to a new module,
  `zipline.pipeline.factors.basic`.

- Added `DailyReturns` as a new built-in factor.
